requirements.txt
aiohttp==3.10.8
pydantic==2.8.2
fastapi==0.115.0
uvicorn==0.30.1
uv==0.9.5
httpx==0.27.2
PyJWT>=2.0.0
pytest
pydantic-settings
ollama
python-jose
python-multipart